refactor helpers for dates and timezones
