+++ src/app/components/organisms/events/useGetFormData.ts
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react'
 import { FormData } from './CreateEvent.types'
 
 import { extractTime, getLocaleDate, getLocaleGmt } from '@/helpers/dates'
-import { getTimezone } from '@/helpers/get-time-zone'
+import { getUserTimezone } from '@/helpers/timezones'
 import { getCountryByZone } from '@/services/timezones'
 import { useTimeZoneContext } from '@/app/context/useTimeZoneContext'
 
@@ -17,8 +17,8 @@ export const useGetFormData = () => {
     eventDescription: '',
     image: '',
     combo: '',
-    country: 'CO',
-    timezone: 'America/Bogota',
+    country: 'US',
+    timezone: 'America/New_York',
     gmt: 'Z',
     hashtags: [],
   })
@@ -26,8 +26,8 @@ export const useGetFormData = () => {
   useEffect(() => {
     const setInitialFormData = async () => {
       const currentDate = new Date()
-      const timezone = getTimezone()
-      const countryCode = (await getCountryByZone(timezone)) ?? 'CO'
+      const timezone = getUserTimezone()
+      const countryCode = (await getCountryByZone(timezone)) ?? 'US'
 
       const initValue = {
         time: extractTime(currentDate),
